+++ lib/phoenix/digester.ex
@@ -28,7 +28,10 @@ defmodule Phoenix.Digester do
       digested_files = Enum.map(files, &digested_contents(&1, latest, with_vsn?))
 
       save_manifest(digested_files, latest, digests, output_path)
-      Enum.each(digested_files, &write_to_disk(&1, output_path))
+
+      digested_files
+      |> Task.async_stream(&write_to_disk(&1, output_path), ordered: false, timeout: :infinity)
+      |> Stream.run()
     else
       {:error, :invalid_path}
     end
@@ -166,15 +169,22 @@ defmodule Phoenix.Digester do
     Enum.each(compressors, fn compressor ->
       [file_extension | _] = compressor.file_extensions
 
-      with {:ok, compressed_digested} <-
-             compressor.compress_file(file.digested_filename, file.digested_content) do
+      compressed_digested_result =
+        compressor.compress_file(file.digested_filename, file.digested_content)
+
+      with {:ok, compressed_digested} <- compressed_digested_result do
         File.write!(
           Path.join(path, file.digested_filename <> file_extension),
           compressed_digested
         )
       end
 
-      with {:ok, compressed} <- compressor.compress_file(file.filename, file.content) do
+      compress_result =
+        if file.digested_content == file.content,
+          do: compressed_digested_result,
+          else: compressor.compress_file(file.filename, file.content)
+
+      with {:ok, compressed} <- compress_result do
         File.write!(
           Path.join(path, file.filename <> file_extension),
           compressed